feat(project): Initialize Flutter project
This commit introduces the initial file structure and boilerplate code for
the `smart_ahwa_manager` application, generated using the `flutter create`
command.

It includes the default counter app, along with the necessary
configurations for Android, iOS, web, and desktop platforms (Linux,
macOS, Windows). This serves as the foundational setup for subsequent
development.