docs(readme): Simplify demo video embedding in README
This commit refactors the README file to simplify how the demo video is
embedded. The previous HTML `<video>` tag has been replaced with a
standard Markdown image/video link.

This change improves the readability and maintainability of the README's
source code by using a more concise and conventional syntax. It also
updates the video caption to be more descriptive.